print

+++ src/do_restore.c
@@ -233,9 +233,9 @@ void do_restore(int revision, char *path) {
 	printf("total size(B): %ld\n", jcr.data_size);
 	printf("total time(s): %.3f\n", jcr.total_time / 1000000);
 	printf("throughput(MB/s): %.2f\n",
-			(double) jcr.data_size * 1000000 / (1024 * 1024 * jcr.total_time));
+			jcr.data_size * 1000000 / (1024.0 * 1024 * jcr.total_time));
 	printf("speed factor: %.2f\n",
-			(double) jcr.data_size / (1024 * 1024 * jcr.read_container_num));
+			jcr.data_size / (1024.0 * 1024 * jcr.read_container_num));
 
 	printf("read_recipe_time : %.3fs, %.2fMB/s\n",
 			jcr.read_recipe_time / 1000000,
